requirements.txt
flask
beautifulsoup4
pymongo
flask_sqlalchemy
python-dotenv

